Move DivDecorator to Web namespace
