+++ library/Reporting/Web/DivDecorator.php
@@ -1,7 +1,7 @@
 <?php
 // Icinga Reporting | (c) 2018 Icinga GmbH | GPLv2
 
-namespace Icinga\Module\Reporting;
+namespace Icinga\Module\Reporting\Web;
 
 use ipl\Html\Attributes;
 use ipl\Html\BaseHtmlElement;
